# public/assets/swagger.yaml
definitions:
  model.Category:
    properties:
      category_type:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  model.CreateCategory:
    properties:
      category_type_ref_id:
        type: integer
      name:
        type: string
    required:
    - category_type_ref_id
    - name
    type: object
  model.PaginatedCategories:
    properties:
      categories:
        items:
          $ref: '#/definitions/model.Category'
        type: array
      pagination:
        $ref: '#/definitions/model.Pagination'
    type: object
  model.Pagination:
    properties:
      max_rows:
        type: integer
      page:
        type: integer
      pages:
        items:
          type: integer
        type: array
      row_count:
        type: integer
      total_count:
        type: integer
    type: object
  model.UpdateCategory:
    type: object
info:
  contact: {}
paths:
  /v1/category:
    get:
      consumes:
      - application/json
      description: Returns the categories
      operationId: ListCategories
      parameters:
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: category_name_in
        type: array
      - collectionFormat: csv
        in: query
        items:
          type: integer
        name: category_type_id_in
        type: array
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: category_type_name_in
        type: array
      - collectionFormat: csv
        in: query
        items:
          type: integer
        name: ids_in
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.PaginatedCategories'
        "400":
          description: Bad Request
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal Server Error
          schema:
            items:
              type: string
            type: array
      summary: Get Categories
      tags:
      - CategoryService
    patch:
      consumes:
      - application/json
      description: Update a category
      operationId: UpdateCategory
      parameters:
      - description: Category body
        in: body
        name: filters
        schema:
          $ref: '#/definitions/model.UpdateCategory'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Category'
        "400":
          description: Bad Request
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal Server Error
          schema:
            items:
              type: string
            type: array
      summary: Update Category
      tags:
      - CategoryService
    post:
      consumes:
      - application/json
      description: Create a category
      operationId: CreateCategory
      parameters:
      - description: Category filters
        in: body
        name: filters
        schema:
          $ref: '#/definitions/model.CreateCategory'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Category'
        "400":
          description: Bad Request
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal Server Error
          schema:
            items:
              type: string
            type: array
      summary: Create Category
      tags:
      - CategoryService
swagger: "2.0"